// src/domain/twitch-streamer-info.interface.ts
export interface TwitchStreamerInfoInterface {
  readonly id: string;
  readonly login: string;
  readonly display_name: string;
  readonly type: string;
  readonly broadcaster_type: string;
  readonly description: string;
  readonly profile_image_url: string;
  readonly offline_image_url: string;
  readonly viewer_count: 294;
  readonly created_at: string;
}